[core] Move last variables outside of the models folder
