+++ packages/grid/x-data-grid/src/hooks/features/rows/index.ts
@@ -2,3 +2,4 @@ export * from './gridRowsMetaSelector';
 export * from './gridRowsMetaState';
 export * from './gridRowsSelector';
 export type { GridRowsState } from './gridRowsState';
+export { checkGridRowIdIsValid } from './gridRowsUtils';
