+++ packages/grid/x-data-grid/src/models/colDef/index.ts
@@ -4,4 +4,3 @@ export * from './gridColDef';
 // Other types
 export * from './gridColType';
 export * from './gridColumnTypesRecord';
-export * from './gridDefaultColumnTypes';
