+++ packages/grid/x-data-grid/src/models/gridRows.ts
@@ -97,26 +97,3 @@ export interface GridRowEntry<R extends GridValidRowModel = any> {
  * The function to retrieve the id of a [[GridRowModel]].
  */
 export type GridRowIdGetter<R extends GridValidRowModel = any> = (row: R) => GridRowId;
-
-/**
- * An helper function to check if the id provided is valid.
- * @param {GridRowId} id Id as [[GridRowId]].
- * @param {GridRowModel | Partial<GridRowModel>} row Row as [[GridRowModel]].
- * @param {string} detailErrorMessage A custom error message to display for invalid IDs
- */
-export function checkGridRowIdIsValid(
-  id: GridRowId,
-  row: GridRowModel | Partial<GridRowModel>,
-  detailErrorMessage: string = 'A row was provided without id in the rows prop:',
-) {
-  if (id == null) {
-    throw new Error(
-      [
-        'MUI: The data grid component requires all rows to have a unique `id` property.',
-        'Alternatively, you can use the `getRowId` prop to specify a custom id for each row.',
-        detailErrorMessage,
-        JSON.stringify(row),
-      ].join('\n'),
-    );
-  }
-}
